Support defining event names in constructor.
This commit implements the enhancement requested in #1.

Events that are initialized by a list/tuple of event names,
will have those names stored in __events__, local to that instance only.

Both __events__ and __class__.__events__ will be evaluated, long as they
exist. They can but should not be used together.